+++ src/oadViewer.js
@@ -2,7 +2,6 @@
 
 const vscode = require('vscode');
 const path = require('path');
-const fs = require('fs');
 
 const oadLogger = require('./oadLogger');
 
@@ -19,49 +18,16 @@ class Viewer {
     }
 
     provideTextDocumentContent() {
-        try{
-            var indexPath = path.join(__dirname, "..", "index.html");
-            logger.log('Index.html: ' + indexPath);
-
-            var baseHtml = fs.readFileSync(indexPath).toString('utf-8');
-
-            // if loading in browser
-            //var finalHtml = baseHtml.replace(/{PREVIEW_URI}/gi, "http://localhost:9005");
-            //return finalHtml;
-            
-            //Else Get resource paths
-            var swagger_editor_dist_path = path.join(this.context.extensionPath, 'node_modules', 'swagger-editor-dist');
-
-            var swagger_editor_css = vscode.Uri.file(path.join(swagger_editor_dist_path, 'swagger-editor.css')).with({ scheme: 'vscode-resource' }).toString(true);
-            var swagger_editor_favicon_32_png = vscode.Uri.file(path.join(swagger_editor_dist_path, "favicon-32x32.png")).with({ scheme: 'vscode-resource' }).toString(true);
-            var swagger_editor_favicon_16_png = vscode.Uri.file(path.join(swagger_editor_dist_path, "favicon-16x16.png")).with({ scheme: 'vscode-resource' }).toString(true);
-            var swagger_editor_bundle_js = vscode.Uri.file(path.join(swagger_editor_dist_path, "swagger-editor-bundle.js")).with({ scheme: 'vscode-resource' }).toString(true);
-            var swagger_editor_standalone_preset_js = vscode.Uri.file(path.join(swagger_editor_dist_path, "swagger-editor-standalone-preset.js")).with({ scheme: 'vscode-resource' }).toString(true);
-            
-            var socket_io = vscode.Uri.file(path.join(this.context.extensionPath, "node_modules", "socket.io", "socket.io.js")).with({ scheme: 'vscode-resource' }).toString(true);
-            
-            // Replace resource paths
-            //var finalHtml = baseHtml.replace(/{SERVER_URL}/gi, "http://localhost:9005");
-            // finalHtml = finalHtml.replace("/node_modules/swagger-editor-dist/favicon-32x32.png", swagger_editor_css);
-            // finalHtml = finalHtml.replace("/node_modules/swagger-editor-dist/favicon-16x16.png", swagger_editor_favicon_32_png);
-            // finalHtml = finalHtml.replace("/node_modules/swagger-editor-dist/favicon-16x16.png", swagger_editor_favicon_16_png);
-            // finalHtml = finalHtml.replace("/node_modules/swagger-editor-dist/swagger-editor-bundle.js", swagger_editor_bundle_js);
-            // finalHtml = finalHtml.replace("/node_modules/swagger-editor-dist/swagger-editor-standalone-preset.js", swagger_editor_standalone_preset_js);
-            // finalHtml = finalHtml.replace("/socket.io/socket.io.js", socket_io);
-
-            var finalHtml = `
-			<html>
-				<body style="margin:0px;padding:0px;background:#fafafa;">
-					<div style="position:fixed;height:100%;width:100%;">
-					<iframe src="http://localhost:${this.port}" frameborder="0" style="overflow:hidden;height:100%;width:100%" height="100%" width="100%"></iframe>
-					</div>
-                </body>
-            <html>`;
-		
-            return finalHtml;
-        } catch (e){
-            logger.log(e);
-        }
+        var html = `
+        <html>
+            <body style="margin:0px;padding:0px;background:#fafafa;">
+                <div style="position:fixed;height:100%;width:100%;">
+                <iframe src="http://localhost:${this.port}" frameborder="0" style="overflow:hidden;height:100%;width:100%" height="100%" width="100%"></iframe>
+                </div>
+            </body>
+        <html>`;
+    
+        return html;
     }
 
     display() {
@@ -78,13 +44,12 @@ class Viewer {
             vscode.ViewColumn.Two,
             {
                 enableScripts: true,
-                retainContextWhenHidden: true,
-                //localResourceRoots: [vscode.Uri.file(path.join(this.context.extensionPath, 'node_modules'))]
+                retainContextWhenHidden: true
             } 
         );
         
         this.currentPanel.webview.html = this.provideTextDocumentContent();
-        //this.currentPanel.onDidDispose(() => this.dispose(), null, null);
+        this.currentPanel.onDidDispose(() => this.dispose(), null, null);
     }
 
     register() {